requirements.txt
SQLAlchemy==2.0.36
python-dotenv==1.0.1
asyncpg==0.30.0
tenacity==9.0.0
python-fasthtml>=0.1.0
aiosqlite==0.20.0